move resolve-user-path to server-utils
